+++ functions/Add-Border.ps1
@@ -16,10 +16,12 @@ Function Add-Border {
 
         [Parameter(Position = 0, Mandatory, ParameterSetName = 'block')]
         [ValidateNotNullOrEmpty()]
+        [Alias("tb")]
         [string[]]$TextBlock,
 
         # The character to use for the border. It must be a single character.
         [ValidateNotNullOrEmpty()]
+        [alias("border")]
         [string]$Character = "*",
 
         # add blank lines before and after text
@@ -28,10 +30,10 @@ Function Add-Border {
         # insert X number of tabs
         [int]$Tab = 0,
 
-        [Parameter(HelpMessage ="Enter an ANSI escape sequence to color the border characters." )]
+        [Parameter(HelpMessage = "Enter an ANSI escape sequence to color the border characters." )]
         [string]$ANSIBorder,
 
-        [Parameter(HelpMessage ="Enter an ANSI escape sequence to color the text." )]
+        [Parameter(HelpMessage = "Enter an ANSI escape sequence to color the text." )]
         [string]$ANSIText
     )
 
@@ -50,38 +52,65 @@ Function Add-Border {
         }
         if ($PSBoundParameters.ContainsKey("ANSIBorder")) {
             Write-Detail "Using an ANSI border Color" -Prefix Begin | Write-Verbose
-            $Character = "{0}{1}{2}" -f $PSBoundParameters.ANSIBorder,$Character,$ansiClear
+            $Character = "{0}{1}{2}" -f $PSBoundParameters.ANSIBorder, $Character, $ansiClear
         }
 
-     } #begin
+        #define regex expressions to detect ANSI escapes. Need to subtract their
+        #length from the string if used. Issue #79
+
+        [regex]$ansiopen = "$([char]0x1b)\[\d+[\d;]+m"
+        [regex]$ansiend = "$([char]0x1b)\[0m"
+
+    } #begin
 
     Process {
 
         if ($pscmdlet.ParameterSetName -eq 'single') {
             Write-Detail "Processing '$text'" -Prefix PROCESS | Write-Verbose
             #get length of text
-            $len = $text.Length
+            $adjust = 0
+            if ($ansiopen.IsMatch($text)) {
+                $adjust += ($ansiopen.matches($text) | Measure-Object length -sum).sum
+                $adjust += ($ansiend.matches($text) | Measure-Object length -sum).sum
+                Write-Detail "Adjusting text length by $adjust." -Prefix PROCESS | Write-Verbose
+            }
+
+            $len = $text.Length - $adjust
             if ($PSBoundParameters.ContainsKey("ANSIText")) {
                 Write-Detail "Using an ANSIText color" -Prefix PROCESS | Write-Verbose
-                $text = "{0}{1}{2}" -f $PSBoundParameters.ANSIText,$text,$AnsiClear
+                $text = "{0}{1}{2}" -f $PSBoundParameters.ANSIText, $text, $AnsiClear
             }
         }
         else {
             Write-Detail "Processing text block" -Prefix PROCESS | Write-Verbose
-            if ($PSBoundparameters.ContainsKey("ANSIText")) {
-                    Write-Detail "Using ANSIText for the block" -prefix PROCESS | Write-Verbose
-                $txtarray = $textblock.split("`n").Trim() | foreach-object {"{0}{1}{2}" -f $PSBoundParameters.ANSIText,$_,$AnsiClear}
+            #test if text block is already using ANSI
+            if ($ansiopen.IsMatch($TextBlock)) {
+                Write-Detail "Text block contains ANSI sequences" -Prefix PROCESS | Write-Verbose
+                $txtarray | ForEach-Object -begin {$tempLen = @()} -process {
+                    $adjust = 0
+                    $adjust += ($ansiopen.matches($_) | Measure-Object length -sum).sum
+                    $adjust += ($ansiend.matches($_) | Measure-Object length -sum).sum
+                    Write-Detail "Length detected as $($_.length)" -Prefix PROCESS | Write-Verbose
+                    Write-Detail "Adding adjustment $adjust" -Prefix PROCESS | Write-Verbose
+                    $tempLen += $_.length - $adjust
+                }
+                $len = $tempLen | Sort-Object -Descending | Select-Object -first 1
 
-                $len = ($txtarray | Sort-Object -property length -Descending | Select-Object -first 1 -expandProperty length) - ($psboundparameters.ANSIText.length+4)
+            }
+            elseif ($PSBoundparameters.ContainsKey("ANSIText")) {
+                Write-Detail "Using ANSIText for the block" -prefix PROCESS | Write-Verbose
+                $txtarray = $textblock.split("`n").Trim() | ForEach-Object {"{0}{1}{2}" -f $PSBoundParameters.ANSIText, $_, $AnsiClear}
+                $len = ($txtarray | Sort-Object -property length -Descending | Select-Object -first 1 -expandProperty length) - ($psboundparameters.ANSIText.length + 4)
             }
             else {
-                 $txtarray = $textblock.split("`n").Trim()
-                 $len = $txtarray | Sort-Object -property length -Descending | Select-Object -first 1 -expandProperty length
+                Write-Detail "Processing simple text block" -prefix PROCESS | Write-Verbose
+                $txtarray = $textblock.split("`n").Trim()
+                $len = $txtarray | Sort-Object -property length -Descending | Select-Object -first 1 -expandProperty length
             }
             Write-Detail "Added $($txtarray.count) text block elements" -Prefix PROCESS | Write-Verbose
-          }
+        }
 
-        Write-Detail "Using a length of $len" | Write-Verbose
+        Write-Detail "Using a length of $len" -Prefix PROCESS | Write-Verbose
         #define a horizontal line
         $hzline = $Character * ($len + 4)
 
@@ -93,15 +122,28 @@ Function Add-Border {
             Write-Detail "Defining Textblock body" -prefix PROCESS | Write-Verbose
             [string[]]$body = $null
             foreach ($item in $txtarray) {
-                    Write-Detail $item -Prefix PROCESS | write-Verbose
-                 if ($PSBoundparameters.ContainsKey("ANSIText")) {
-                    #adjust the padding length to take the ANSI value into account
-                    $adjust = $len+($psboundparameters.ANSIText.length+4)
+                if ($item) {
+                    Write-Detail "$item [$($item.length)]" -Prefix PROCESS | Write-Verbose
+                }
+                else {
+                    Write-Detail "detected blank line" -Prefix PROCESS | Write-Verbose
+                }
+                if ($ansiopen.IsMatch($item)) {
+                    $adjust = $len
+                    $adjust += ($ansiopen.matches($item) | Measure-Object length -sum).sum
+                    $adjust += ($ansiend.matches($item) | Measure-Object length -sum).sum
                     Write-Detail "Adjusting length to $adjust" -prefix PROCESS | Write-Verbose
+                    $body += "$tabs$Character $(($item).PadRight($adjust)) $Character`r"
+
+                }
+                elseif ($PSBoundparameters.ContainsKey("ANSIText")) {
+                    #adjust the padding length to take the ANSI value into account
+                    $adjust = $len + ($psboundparameters.ANSIText.length + 4)
+                    Write-Detail "Adjusting length to $adjust" -prefix PROCESS | Write-Verbose
+
                     $body += "$tabs$Character $(($item).PadRight($adjust)) $Character`r"
                 }
                 else {
-
                     $body += "$tabs$Character $(($item).PadRight($len)) $Character`r"
                 }
             } #foreach item in txtarray