+++ samples/DiskData.ps1
@@ -1,4 +1,5 @@
 #requires -module PSScriptTools
+#requires -version 5.1
 
 Function Get-DiskData {
 [cmdletbinding()]
@@ -9,15 +10,15 @@ Param(
    )
 
    Begin {
-        Write-Detail "Starting $($myinvocation.MyCommand)" -Prefix begin -NoDate | Write-Verbose
+        Write-Detail "Starting $($myinvocation.MyCommand)" -Prefix BEGIN  | Write-Verbose
    } #begin
 
    Process {
-        Write-Detail "Processing $($computername.toUpper())" -Prefix process -NoDate | Write-Verbose
+        Write-Detail "Processing $($computername.toUpper())" -Prefix PROCESS  | Write-Verbose
         Try {
             $data = Get-CimInstance -Class Win32_logicaldisk -Filter "DriveType=3" -ComputerName $Computername -ErrorAction Stop
             $data | Foreach-Object {
-                Write-Detail "Calculating PctFree for $($_.DeviceID)" -prefix process -NoDate | Write-Verbose
+                Write-Detail "Calculating PctFree for $($_.DeviceID)" -prefix PROCESS  | Write-Verbose
                 $_ | Add-Member -MemberType ScriptProperty -Name PctFree -value { Format-Percent -value $this.freespace -total $this.size -decimal 2} -force
             }
             $data
@@ -28,7 +29,7 @@ Param(
    } #process
 
    End {
-        Write-Detail "Ending $($myinvocation.MyCommand)" -Prefix end -NoDate | Write-Verbose
+        Write-Detail "Ending $($myinvocation.MyCommand)" -Prefix END  | Write-Verbose
    } #end
 
 }